+++ CHANGELOG.md
@@ -9,9 +9,8 @@
 *   Bug Fixes:
     *   Fixed starring episode from full-screen player does not prevent episode from being archived
         ([#1735](https://github.com/Automattic/pocket-casts-android/pull/1735))
-*   New features:
-    *   Display episode artwork from podcast feed on Mini-player, Notification & Widget.
-        ([#1658](https://github.com/Automattic/pocket-casts-android/pull/1558))
+    *   Fixed "Hide playback notification on pause" setting not representing its state correctly
+        ([#1769](https://github.com/Automattic/pocket-casts-android/pull/1769))
 
 7.56
 -----